Fix: Versioning and Kobo normalization

+++ src/main/java/dev/slethware/walletservice/models/dtos/response/BalanceResponse.java
@@ -5,13 +5,11 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class BalanceResponse {
-    private BigDecimal balance;
+    private Long balance;
     private String walletNumber;
 }
